+++ src/test/java/de/alpharogroup/gson/type/adapters/BigDecimalMoneyScaledAdapterTest.java
@@ -31,7 +31,6 @@ import com.google.gson.stream.JsonWriter;
 import de.alpharogroup.file.search.PathFinder;
 import de.alpharogroup.gson.JsonFileToObjectExtensions;
 import de.alpharogroup.gson.Signin;
-import de.alpharogroup.gson.factory.TypeFactory;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;